excel-to-pdf-converter/backend/requirements.txt
fastapi
uvicorn
Pillow
python-multipart
Flask
werkzeug
docx2pdf
pandas
xlsx2html
pdfkit
wkhtmltopdf